% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIDfinder.R
\name{build_rcid}
\alias{build_rcid}
\title{function identifying the RCID to be assigned in the legacy data frame}
\usage{
build_rcid(df_leg, df_rcid)
}
\value{
the input legacy data frame after updates to the RCID column, and the updated table for the rcid too
}
\description{
function identifying the RCID to be assigned in the legacy data frame
}
