% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIDfinder.R
\name{extract_year}
\alias{extract_year}
\title{function to split the legacy data Date column (yyyy-mm) in two columns containing year and month, to be added to the df}
\usage{
extract_year(df_rc)
}
\value{
a new data frame, copying over the input one and adding the new columns
}
\description{
function to split the legacy data Date column (yyyy-mm) in two columns containing year and month, to be added to the df
}
