% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIDfinder.R
\name{is_old_cycle}
\alias{is_old_cycle}
\title{function identifying whether a cycle belongs to the RCID era, or predates it, and needs therefore to be classified on the year-based RCID}
\usage{
is_old_cycle(df)
}
\value{
the input data frame after manipulation
}
\description{
function identifying whether a cycle belongs to the RCID era, or predates it, and needs therefore to be classified on the year-based RCID
}
