% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_rcid_table.R
\name{map_country_codes}
\alias{map_country_codes}
\title{function to create a mapping of the country and country codes}
\usage{
map_country_codes(df)
}
\value{
the data frame for the RCID with two extra columns
}
\description{
function to create a mapping of the country and country codes
}
