% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_rcid_table.R
\name{map_cycle_year}
\alias{map_cycle_year}
\title{function to map the cycle numbers to the years}
\usage{
map_cycle_year(df)
}
\value{
df with extra column with year definition
}
\description{
function to map the cycle numbers to the years
}
